backend/yourluxuryhome/requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2

# REST Framework
djangorestframework==3.14.0
markdown==3.5
django-filter==24.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# API Documentation
drf-yasg==1.21.7

# CORS
django-cors-headers==4.3.1

# Nested Routers
drf-nested-routers==0.93.4

# Image Processing
Pillow==10.4.0

# Environment Variables
python-dotenv==1.0.1

# Date/Time Handling
pytz==2024.1

# Email Verification
djoser==2.2.2
django-templated-mail==1.1.1

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
